requirements.txt
celery
redis
pyyaml
pymodbus==3.6.8